Merge branch 'master' into stable

Conflicts:
	libs/qa_lib_tblib/qa_lib_tblib/css/screen.css
	tools/qa_hamsta/qa_hamsta.spec
	tools/qa_hamsta/qa_hamsta/Slave/slave_diplom.pl
	tools/qa_hamsta/qa_hamsta/frontend/globals.php
	tools/qa_qadb/qa_qadb/frontend/login.php
	tools/qa_qadb/qa_qadb/frontend/qadb.php
	tools/qa_qadb/qa_qadb/frontend/submission.php
	tools/qa_qadb/qa_qadb/frontend/waiver.php
	tools/qa_tools/qa_tools/setupUIAutomationtest

+++ tools/qa_tools/qa_tools/reinstall.pl
@@ -39,6 +39,7 @@ use Getopt::Long qw(:config no_ignore_case no_getopt_compat);
 
 use qaconfig;
 use install_functions;
+use log;
 
 $ENV{'LC_ALL'}='en_US';
 
@@ -47,6 +48,7 @@ our $args={
 	rootfstype=>'ext3',
 	defaultboot=>'',
 	setupfordesktoptest=>'',
+        kexecboot=>'',
 };
 
 $args->{'newvm'} = ( $0 =~ /newvm(?:\.\w*)?$/ );
@@ -67,8 +69,10 @@ our $options = [
 	['P','repartitiondisk','1..100','Repartition entire disk, use so many percent of it','When you specify a number between 1-100 here, all partitions will be deleted, then root, swap, /abuild and /boot (if existing) will be recreated. These four partitions together should use the given percentage of the disk size. In the rest of the disk, you can create new partitions for testing.'],
 	['B','setup_bridge','','Configure the system network interface to bridge','Create a bridging pseudo-device associated with primary ethernet adapter. Happens automatically for virtualization hosts.'],
 	['D','setupfordesktoptest','','Configure the system to run desktop tests','Prepares for automated desktop testing. Turns off desktop access control for X, adds accessibility features, plus more.'],
+        ['k','kexecboot','','kexec load install','Use Kexec to load install'],
 	['?','help','','Print help message',''],
 	['N','manual','','Generate manual page','Used to generate this manual page and keep it up-to-date.'],
+	['Z','timezone','timezone','time zone','A time zone for SUT.'],
 
 	### Update options ###
 	['O','opensuse_update','','Add OpenSuSE update repository',"Similar to '-s', but you don't have to type in the URL, it will be generated."],
@@ -270,6 +274,11 @@ unless ($args->{'winvm'}) {
 
 		$ay_xml = 'autoinst_'.$args->{'hostname'} . ($args->{'newvm'} ? "_vm_$$" : '') . '.xml';
 		&command( "$tooldir/modify_xml.pl -m '$modfile' '$profile' '$mountpoint/autoinst/$ay_xml'" );
+		if ($args->{'timezone'}) {
+			my ($continent, $city) = split "_", $args->{'timezone'};
+			&command (" sed -i 's/Europe/$continent/' $mountpoint/autoinst/$ay_xml");
+			&command (" sed -i 's/Prague/$city/' $mountpoint/autoinst/$ay_xml");
+		}
 		$ay_xml = $qaconf{'install_profile_url_base'} . '/' . $ay_xml;
 
 		&command( "umount $mountpoint" );
@@ -332,13 +341,29 @@ if( $args->{'newvm'} )	{
 	system ("touch /var/lib/hamsta/stats_changed");
 	exit $ret;
 } else { # reinstall / upgrade
-	my $cmdline = "$aytool ".$args->{'source'}.' ';
+	my $boottype = "bootloader";
+	if ( $args->{'kexecboot'} ) {
+		$boottype = "kexecboot";
+	}
+	my $cmdline = "$aytool ".$args->{'source'}.' '."$boottype".' ';
 	$cmdline .= "autoupgrade=1 " if $args->{'upgrade'};
 	$cmdline .= "autoyast=$ay_xml install=".$args->{'source'};
 	$cmdline .= ' '.$args->{'installoptions'} if defined $args->{'installoptions'};
-	print $cmdline . "\n";
 	&command($cmdline);
-	&command( "sleep 2" );
-	&command( "reboot" );
-	exit -1;
+        &command( "sleep 2" );
+
+	if ( "$boottype" eq "bootloader" ) {
+        	&command( "reboot" );
+		exit -1;
+	} else {
+		my $pid = fork();
+		if ( $pid > 0 ) {
+			$SIG{CHLD} = 'IGNORE';
+			exit 0;
+		} else {
+			&log(LOG_RETURN, "$? (".$cmdline.')');
+       			system("sleep 20");
+			exec("/sbin/kexec -e >/dev/null");
+		}
+	}
 }
